chore: rename to aws
