+++ src/cores/aws/aws-s3-service.ts
