+++ src/cores/aws/aws-sns-service.ts
